requirements.txt
async-timeout==5.0.1 ; python_version >= "3.11" and python_full_version < "3.11.3"
blinker==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
cachelib==0.13.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.2.1 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
compressed-rtf==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
contourpy==1.3.2 ; python_version >= "3.11" and python_version < "4.0"
cycler==0.12.1 ; python_version >= "3.11" and python_version < "4.0"
dotenv==0.9.9 ; python_version >= "3.11" and python_version < "4.0"
flask-session==0.8.0 ; python_version >= "3.11" and python_version < "4.0"
flask==3.1.1 ; python_version >= "3.11" and python_version < "4.0"
fonttools==4.58.4 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
kiwisolver==1.4.8 ; python_version >= "3.11" and python_version < "4.0"
lxml==5.4.0 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
matplotlib==3.10.3 ; python_version >= "3.11" and python_version < "4.0"
msgspec==0.19.0 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.3.1 ; python_version >= "3.11" and python_version < "4.0"
packaging==25.0 ; python_version >= "3.11" and python_version < "4.0"
pillow==11.2.1 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.2.3 ; python_version >= "3.11" and python_version < "4.0"
pypdf2==3.0.1 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-docx==1.2.0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.11" and python_version < "4.0"
redis==6.2.0 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.14.0 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
wordcloud==1.9.4 ; python_version >= "3.11" and python_version < "4.0"
